pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["ddertaliss <ddertaliss@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
django-bootstrap5 = "^24.2"
django-i18n = "^1.0.5"
django-modeltranslation = "^0.19.7"
flake8 = "^7.1.1"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.2"
django-filter = "^24.3"
rollbar = "^0.16.3"
coverage = "^7.6.1"
pip = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
